+++ src/main/java/com/example/formacionBitboxer2/service/IReduccionService.java
@@ -5,7 +5,7 @@ import com.example.formacionBitboxer2.entities.Reduccion;
 
 public interface IReduccionService {
     public void asociarArticulo(int codigo, ReduccionDTO reduccionDTO);
-    public ReduccionDTO obtenerPorCodigoreduccion(Integer codigo);
+    public ReduccionDTO obtenerPorCodigoreduccion(Long codigo);
     public void guardarReduccion(Reduccion reduccion);
 
 }
